Switch Bitwise from veneer to private macro interface
