+++ src/main/scala-2/chisel3/util/BitwiseIntf.scala
@@ -10,18 +10,7 @@ import chisel3.experimental.SourceInfo
 import chisel3.internal.sourceinfo.SourceInfoTransform
 import scala.language.experimental.macros
 
-/** Creates repetitions of each bit of the input in order.
-  *
-  * @example {{{
-  * FillInterleaved(2, "b1 0 0 0".U)  // equivalent to "b11 00 00 00".U
-  * FillInterleaved(2, "b1 0 0 1".U)  // equivalent to "b11 00 00 11".U
-  * FillInterleaved(2, myUIntWire)  // dynamic interleaved fill
-  *
-  * FillInterleaved(2, Seq(false.B, false.B, false.B, true.B))  // equivalent to "b11 00 00 00".U
-  * FillInterleaved(2, Seq(true.B, false.B, false.B, true.B))  // equivalent to "b11 00 00 11".U
-  * }}}
-  */
-object FillInterleaved extends FillInterleavedImpl {
+private[chisel3] trait FillInterleaved$Intf { self: FillInterleaved.type =>
 
   /** Creates n repetitions of each bit of x in order.
     *
@@ -42,18 +31,7 @@ object FillInterleaved extends FillInterleavedImpl {
   def do_apply(n: Int, in: Seq[Bool])(implicit sourceInfo: SourceInfo): UInt = _applyImpl(n, in)
 }
 
-/** Returns the number of bits set (value is 1 or true) in the input signal.
-  *
-  * @example {{{
-  * PopCount(Seq(true.B, false.B, true.B, true.B))  // evaluates to 3.U
-  * PopCount(Seq(false.B, false.B, true.B, false.B))  // evaluates to 1.U
-  *
-  * PopCount("b1011".U)  // evaluates to 3.U
-  * PopCount("b0010".U)  // evaluates to 1.U
-  * PopCount(myUIntWire)  // dynamic count
-  * }}}
-  */
-object PopCount extends PopCountImpl {
+private[chisel3] trait PopCount$Intf { self: PopCount.type =>
 
   def apply(in: Iterable[Bool]): UInt = macro SourceInfoTransform.inArg
 
@@ -66,15 +44,7 @@ object PopCount extends PopCountImpl {
   def do_apply(in: Bits)(implicit sourceInfo: SourceInfo): UInt = _applyImpl(in)
 }
 
-/** Create repetitions of the input using a tree fanout topology.
-  *
-  * @example {{{
-  * Fill(2, "b1000".U)  // equivalent to "b1000 1000".U
-  * Fill(2, "b1001".U)  // equivalent to "b1001 1001".U
-  * Fill(2, myUIntWire)  // dynamic fill
-  * }}}
-  */
-object Fill extends FillImpl {
+private[chisel3] trait Fill$Intf { self: Fill.type =>
 
   /** Create n repetitions of x using a tree fanout topology.
     *
@@ -87,15 +57,7 @@ object Fill extends FillImpl {
   def do_apply(n: Int, x: UInt)(implicit sourceInfo: SourceInfo): UInt = _applyImpl(n, x)
 }
 
-/** Returns the input in bit-reversed order. Useful for little/big-endian conversion.
-  *
-  * @example {{{
-  * Reverse("b1101".U)  // equivalent to "b1011".U
-  * Reverse("b1101".U(8.W))  // equivalent to "b10110000".U
-  * Reverse(myUIntWire)  // dynamic reverse
-  * }}}
-  */
-object Reverse extends ReverseImpl {
+private[chisel3] trait Reverse$Intf { self: Reverse.type =>
 
   def apply(in: UInt): UInt = macro SourceInfoTransform.inArg
